requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
requests==2.32.3
gspread==6.1.2
google-auth==2.30.0
openpyxl==3.1.4
streamlit>=1.35.0
plotly>=5.20.0
anthropic>=0.25.0
